bugs fixed

+++ FoodOrderingApplicationFigma/Services/Service-Interfaces/ICustomerService.cs
@@ -5,9 +5,9 @@ namespace FoodOrderingApplicationFigma.Services.Service_Interfaces
     public interface ICustomerService
     {
         Task<IEnumerable<GetCustomerDTO>> GetAllCustomers();
-        Task<GetCustomerDTO?> GetCustomerById(int id);
+        Task<GetCustomerDTO?> GetCustomerById(long id);
         Task<GetCustomerDTO> CreateCustomer(CreateCustomerDTO createCustomerDTO);
         Task<GetCustomerDTO?> UpdateCustomer(UpdateCustomerDTO updateCustomerDTO);
-        Task<GetCustomerDTO> DeleteCustomer(int id);
+        Task<GetCustomerDTO> DeleteCustomer(long id);
     }
 }
